server/requirements.txt
flask_login
flask_socketio
flask_wtf
email_validator
blinker
psycopg2-binary
python-dotenv
